feat: Root Reducer tests added
