+++ src/services/slices/user/userSlice.test.ts
@@ -1,0 +1,12 @@
+import { registerUser, loginUser, logout, checkUserAuth, TUserState } from './userSlice';
+import userReducer from './userSlice';
+
+
+describe('User Slice Tests', () => {
+    const initialState: TUserState = {
+        isAuthChecked: false,
+        user: null
+    };
+
+
+})
