+++ src/services/slices/user/userSlice.ts
@@ -9,10 +9,10 @@ import {
   logoutApi,
   registerUserApi,
   updateUserApi
-} from '@api';
+} from '../../../utils/burger-api';
 import { TUser } from '@utils-types';
 
-type TUserState = {
+export type TUserState = {
   isAuthChecked: boolean;
   user: TUser | null;
 };
