+++ src/services/store.ts
@@ -5,7 +5,7 @@ import {
   useDispatch as dispatchHook,
   useSelector as selectorHook
 } from 'react-redux';
-import { rootReducer } from './rootReducer';
+import { rootReducer } from './rootReducer/rootReducer';
 
 const store = configureStore({
   reducer: rootReducer,
